+++ app/page.tsx
@@ -1,0 +1,5 @@
+import WhatsOn from "./components/WhatsOn";
+
+export default function Home() {
+  return <WhatsOn />;
+}